+++ docs/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## TODO
 # Refine entity embed for guide markdown
 
+## 2026-02-20
+- Fluxer login option added with updates to the frontend to reflect this change
+- Profile page now allows user to link or unlink their Discord/Fluxer
+
 ## 2026-02-10
 - Entity display media management for admin dashboard
 - Media preview fixed for admin dashboard